"controller asistencia"

+++ controllers/AsistenciaController.php
@@ -3,10 +3,13 @@
 namespace app\controllers;
 
 use app\models\Asistencia;
+use yii\data\ActiveDataProvider;
 use app\models\AsistenciaSearch;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
+
+use app\models\Permiso;
 
 /**
  * AsistenciaController implements the CRUD actions for Asistencia model.
@@ -38,15 +41,29 @@ class AsistenciaController extends Controller
      */
     public function actionIndex()
     {
-        $searchModel = new AsistenciaSearch();
-        $dataProvider = $searchModel->search($this->request->queryParams);
+        if (!Permiso::accion('asistencia', 'index')) {
+            return $this->render('/site/error', [
+                'name' => 'Permiso denegado',
+                'message' => 'No tiene permiso para realizar esta función, verifique con el administrador de sistemas.'
+            ]);
+        }
+        $dataProvider = new ActiveDataProvider([
+            'query' => Asistencia::find(),
+            'pagination' => [
+                'pageSize' => 25
+            ],
+           /* 'sort' => [
+                'defaultOrder' => [
+                    'id' => SORT_DESC,
+                ]
+            ],
+            */
+        ]);
 
         return $this->render('index', [
-            'searchModel' => $searchModel,
             'dataProvider' => $dataProvider,
         ]);
     }
-
     /**
      * Displays a single Asistencia model.
      * @param int $id ID
@@ -55,6 +72,13 @@ class AsistenciaController extends Controller
      */
     public function actionView($id)
     {
+        if (!Permiso::accion('asistencia', 'view')) {
+            return $this->render('/site/error', [
+                'name' => 'Permiso denegado',
+                'message' => 'No tiene permiso para realizar esta función, verifique con el administrador de sistemas.'
+            ]);
+        }
+
         return $this->render('view', [
             'model' => $this->findModel($id),
         ]);
@@ -67,6 +91,13 @@ class AsistenciaController extends Controller
      */
     public function actionCreate()
     {
+        if (!Permiso::accion('asistencia', 'create')) {
+            return $this->render('/site/error', [
+                'name' => 'Permiso denegado',
+                'message' => 'No tiene permiso para realizar esta función, verifique con el administrador de sistemas.'
+            ]);
+        }
+
         $model = new Asistencia();
 
         if ($this->request->isPost) {
@@ -101,6 +132,20 @@ class AsistenciaController extends Controller
             'model' => $model,
         ]);
     }
+        public function actionUpdateEstatus($id)
+    {
+        if (!Permiso::accion('asistencia', 'update-estatus')) {
+            return $this->render('/site/error', [
+                'name' => 'Permiso denegado',
+                'message' => 'No tiene permiso para realizar esta función, verifique con el administrador de sistemas.'
+            ]);
+        }
+
+        $model = $this->findModel($id);
+        if ($this->request->isPost && $model->load($this->request->post()) && $model->save()) {
+        }
+        return $this->redirect(['index']);
+    }
 
     /**
      * Deletes an existing Asistencia model.
@@ -111,11 +156,17 @@ class AsistenciaController extends Controller
      */
     public function actionDelete($id)
     {
+        if (!Permiso::accion('asistencia', 'delete')) {
+            return $this->render('/site/error', [
+                'name' => 'Permiso denegado',
+                'message' => 'No tiene permiso para realizar esta función, verifique con el administrador de sistemas.'
+            ]);
+        }
+
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
     }
-
     /**
      * Finds the Asistencia model based on its primary key value.
      * If the model is not found, a 404 HTTP exception will be thrown.
